requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.5.2
appdirs==1.4.4
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
bentoml==1.3.5
cattrs==23.1.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
circus==0.19.0
click==8.1.8
click-option-group==0.5.7
cloudpickle==3.1.1
cryptography==44.0.2
deepmerge==2.0
Deprecated==1.2.18
eval-type-backport==0.2.2
exceptiongroup==1.2.2
frozenlist==1.5.0
fs==2.4.16
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-ws==0.6.2
idna==3.10
importlib-metadata==6.11.0
inflection==0.5.1
inquirerpy==0.3.4
jinja2==3.1.6
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
numpy==1.24.4
nvidia-ml-py==11.525.150
opentelemetry-api==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
packaging==24.2
pandas==2.0.3
pathspec==0.12.1
pfzy==0.3.4
pip-requirements-parser==32.0.1
prometheus-client==0.21.1
prompt-toolkit==3.0.50
propcache==0.2.0
psutil==7.0.0
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
requests==2.32.3
rich==13.9.4
schema==0.7.7
scikit-learn==1.3.2
scipy==1.10.1
simple-di==0.1.5
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
threadpoolctl==3.5.0
tomli==2.2.1
tomli-w==1.0.0
tornado==6.4.2
typing-extensions==4.13.0
tzdata==2025.2
urllib3==2.2.3
uv==0.6.10
uvicorn==0.33.0
watchfiles==0.24.0
wcwidth==0.2.13
wrapt==1.17.2
wsproto==1.2.0
yarl==1.15.2
zipp==3.20.2
